move inspect name to separate component
