+++ apps/client/src/play/ui/editor/pages/Inspect/InspectPage.tsx
@@ -1,9 +1,7 @@
-import { editNode } from "@/src/play/actions/editNode";
-import TextFieldDark from "@/src/ui/TextFieldDark";
-
 import { useTreeValue } from "../../hooks/useTreeValue";
 import { getDisplayName } from "../../utils/getDisplayName";
 import PanelPage from "../PanelPage";
+import { Name } from "./Name";
 import Physics from "./Physics";
 import Transform from "./Transform";
 
@@ -13,29 +11,12 @@ interface Props {
 
 export default function InspectPage({ id }: Props) {
   const name = useTreeValue(id, "name");
-  const locked = useTreeValue(id, "locked");
 
   const displayName = getDisplayName(name, id);
 
   return (
     <PanelPage title={displayName}>
-      <TextFieldDark
-        label="Name"
-        value={name}
-        disabled={locked}
-        placeholder={displayName}
-        onChange={(e) => {
-          if (!name) {
-            return;
-          }
-
-          editNode({
-            name: e.target.value,
-            target: name,
-          });
-        }}
-      />
-
+      <Name id={id} />
       <Transform id={id} />
       <Physics id={id} />
     </PanelPage>
